requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
django-debug-toolbar==3.8.1
django-filter==22.1
Django==4.1.6
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==23.0
pip==23.0
python-gettext==4.1
pytz==2022.7
requests==2.28.2
ruamel.yaml==0.17.21
sentry-sdk==1.15.0
setuptools==65.5.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
